third_party/WebKit/LayoutTests/svg/dom/svglist-exception-on-out-bounds-error-expected.txt

Tests that out of bounds accesses of SVGList correctly throw exceptions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (4294967295) is greater than the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (4294967295) is greater than the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (4294967295) is greater than the maximum bound (0)..
PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (1) is greater than the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (1) is greater than the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (1) is greater than the maximum bound (0)..
PASS svgList.getItem(index) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Failed to execute 'replaceItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS svgList.replaceItem(null, index) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGPathSegList': parameter 1 is not of type 'SVGPathSeg'..
PASS svgList.removeItem(index) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGPathSegList': The index provided (0) is greater than or equal to the maximum bound (0)..
PASS successfullyParsed is true

TEST COMPLETE
